Update README.md
This system makes the employer, administrator, or HR work easier and more convenient with a company. It ensures structural and professional approach to employee management within a company. 
Administrators and employees can seamlessly interact with the system ensuring improved operational efficiency and transparency in a workplace and in compliance to various SDGs

The system has two user roles; 
1. The Administrator who has the privilege to access the list of all employees, approval and decline employee leave applications, add users, 
2. Employee - can create account, login to the system, apply (create) leave, update details

the system has a responsive layout ensuring its suitability and usability across all devices